XMLRPC server
Adding a new method to activate the manage module. This method will
fail if the request does not provide a valid secret. The secret is
generated at the end of the connection authorization process. The
secret is valid for up to 24 hours. It will allow our users who decide
to purchase a premium plan to activate the manage module, so they
can configure their premium settings remotely.